refactor: split upload controller by type and add plant growth stage support
